requirements.txt
requests
transformers
huggingface_hub
torch
datasets
accelerate
bitsandbytes
sentencepiece
pyyaml
gradio